// front/src/providers/api-client.generated/model/candidateMessageDto.ts
/**
 * Morgan And Mallet CRM
 * Morgan And Mallet CRM API
 *
 * The version of the OpenAPI document: 1.0
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */
import { UserDto } from './userDto';
import { CandidateDto } from './candidateDto';


export interface CandidateMessageDto { 
    id?: string;
    creationDate?: Date;
    modifDate?: Date;
    content: string;
    candidateId?: string;
    candidate?: CandidateDto;
    seen?: boolean;
    senderId?: string;
    sender?: UserDto;
    senderType?: string;
    archived?: boolean;
}